+++ CHANGELOG.md
@@ -17,7 +17,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 
-- Fixed BART. This implementation had some major bugs in it that caused poor beam search predictions.
+- Fixed BART. This implementation had some major bugs in it that caused poor performance during prediction.
 
 ## [v1.3.0](https://github.com/allenai/allennlp-models/releases/tag/v1.2.2) - 2020-12-15